# perf(proto): reduce Value's size by one-word